app/services/subscriber_influx/.ipynb_checkpoints/requirements-checkpoint.txt
paho-mqtt<2
influxdb